refactor: application 레이어에서 Spring Page 클래스 사용하도록 수정
